# data-simulator-service/docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "独立模拟源管理服务",
    "version": "1.0.0",
    "description": "支持PDXP/UDP和FEP/TCP，多模拟源独立并发。默认端口8085。发送端不自动启停接收服务。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8085",
      "description": "本地独立服务"
    }
  ],
  "paths": {
    "/simulator/sources": {
      "get": {
        "summary": "分页查询模拟源",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "每页数量"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "名称包含查询"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6
              ]
            },
            "description": "0未启动1启动中2运行中3停止中4已停止5已完成6异常"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "pageNum": {
                          "type": "integer",
                          "description": "页码"
                        },
                        "pageSize": {
                          "type": "integer",
                          "description": "每页数量"
                        },
                        "total": {
                          "type": "integer",
                          "description": "总数",
                          "format": "int64"
                        },
                        "records": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Source"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/{id}": {
      "get": {
        "summary": "模拟源详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "模拟源编号"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Source"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/{id}/runs": {
      "get": {
        "summary": "运行历史，允许查询已删除源",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "模拟源编号"
          },
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "pageNum": {
                          "type": "integer",
                          "description": "页码"
                        },
                        "pageSize": {
                          "type": "integer",
                          "description": "每页数量"
                        },
                        "total": {
                          "type": "integer",
                          "description": "总数",
                          "format": "int64"
                        },
                        "records": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Run"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/findData": {
      "get": {
        "summary": "查询中文枚举和协议默认值",
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EnumData"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/create": {
      "post": {
        "summary": "新建配置",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "integer",
                      "description": "新建模拟源编号",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/update": {
      "post": {
        "summary": "完整修改配置，禁止切换协议",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/delete": {
      "post": {
        "summary": "逻辑删除配置",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/simulator/sources/start": {
      "post": {
        "summary": "启动独立发送任务",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "结果说明",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "integer",
                      "description": "本次运行编号",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "名称重复、重复启动或运行中编辑等状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务或数据库异常",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "结果码",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "description": "中文错误说明"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Source": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "修改时必填，新建时禁止填写",
            "format": "int64",
            "minimum": 1
          },
          "sourceName": {
            "type": "string",
            "description": "未删除配置中唯一",
            "maxLength": 100
          },
          "transferProtocol": {
            "type": "integer",
            "description": "2为PDXP，4为FEP",
            "enum": [
              2,
              4
            ]
          },
          "transferType": {
            "type": "integer",
            "description": "PDXP只允许1（UDP），FEP只允许2（TCP）；省略时按协议设置",
            "enum": [
              1,
              2
            ]
          },
          "filePath": {
            "type": "string",
            "description": "服务所在机器上的文件路径",
            "maxLength": 1000
          },
          "targetHost": {
            "type": "string",
            "description": "目标地址",
            "default": "127.0.0.1",
            "maxLength": 100
          },
          "targetPort": {
            "type": "integer",
            "description": "PDXP默认9001，FEP默认19003",
            "minimum": 1,
            "maximum": 65535
          },
          "sendIntervalMillis": {
            "type": "integer",
            "description": "数据单元发送间隔；PDXP默认500，FEP默认10",
            "minimum": 0,
            "maximum": 86400000,
            "format": "int64"
          },
          "loopEnabled": {
            "type": "integer",
            "description": "0否1是；PDXP默认1，FEP默认0",
            "enum": [
              0,
              1
            ]
          },
          "remark": {
            "type": "string",
            "description": "备注",
            "maxLength": 500
          },
          "pdxp": {
            "$ref": "#/components/schemas/PdxpConfig"
          },
          "createTime": {
            "type": "string",
            "description": "服务器本地时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "服务器本地时间",
            "format": "date-time"
          },
          "protocolName": {
            "type": "string",
            "description": "协议名称"
          },
          "transferTypeName": {
            "type": "string",
            "description": "传输方式中文名称"
          },
          "loopEnabledName": {
            "type": "string",
            "description": "是否循环中文名称",
            "enum": [
              "否",
              "是"
            ]
          },
          "latestRunId": {
            "type": "integer",
            "description": "最近运行编号",
            "format": "int64",
            "nullable": true
          },
          "sentUnitCount": {
            "type": "integer",
            "description": "最近运行发送单元数量",
            "format": "int64"
          },
          "runStatus": {
            "type": "string",
            "description": "运行状态",
            "enum": [
              "未启动",
              "启动中",
              "运行中",
              "停止中",
              "已停止",
              "已完成",
              "异常"
            ]
          }
        }
      },
...
      "IdRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "模拟源编号",
            "format": "int64",
            "minimum": 1
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false
      },
...
    }
  }
}